Move all required JS from Example into Emission.
