+++ index.js
@@ -5,6 +5,9 @@ import Containers from './lib/containers';
 import Components from './lib/components';
 import Routes from './lib/relay/routes';
 
+import './lib/relay/config';
+import './lib/app_registry';
+
 export default {
   Containers,
   Components,
